Line Detection Work
Hough transforms(probablistic) implemented however not working completely as expected. Increasing the size of accumulators was expected to produce larger lines however areas of the image as still not being detected. Attempts are inprogress to improve the hough transform output.
Algorithm produced to merge lines with a similar gradient and location in image